+++ model/team.h
@@ -7,13 +7,13 @@
 
 typedef struct Team {
     char name[20];
-    int colour;
+    uint8_t colour;
 
     uint8_t piece_count;
     Piece * pieces;
 } Team;
 
-Team * createTeam(char * name, int colour, Piece * pieces, uint8_t piece_count);
+Team * createTeam(char * name, uint8_t colour, Piece * pieces, uint8_t piece_count);
 void saveTeam(Team * team, FILE * stream);
 Team * loadTeam(FILE * stream);
 void printTeam(Team * team);
